Feat: Create API for campaigns.

+++ solution/microservices/ad_engine/src/domain/schemas/request/time.rs
@@ -6,5 +6,5 @@
 pub struct TimeAdvanceRequest {
     #[schema(example = 1)]
     #[validate(range(min = 0, message = "date must be under or equal 0"))]
-    pub current_date: u16,
+    pub current_date: u32,
 }
